+++ lib/ui/academics/StudentCoursesContentWidget.dart
@@ -135,7 +135,7 @@ class _StudentCoursesContentWidgetState extends State<StudentCoursesContentWidge
     TextStyle(fontFamily: Styles().fontFamilies?.medium, fontSize: 16, color: Styles().colors?.fillColorPrimary);
 
   Widget _buildTermsDropDown() {
-    CourseTerm? currentTerm = StudentCourses().displayTerm;
+    StudentCourseTerm? currentTerm = StudentCourses().displayTerm;
 
     return Semantics(label: currentTerm?.name, hint: "Double tap to select account", button: true, container: true, child:
       DropdownButtonHideUnderline(child:
@@ -153,13 +153,13 @@ class _StudentCoursesContentWidgetState extends State<StudentCoursesContentWidge
   }
 
   List<DropdownMenuItem<String>>? _buildTermDropDownItems() {
-    List<CourseTerm>? terms = StudentCourses().terms;
+    List<StudentCourseTerm>? terms = StudentCourses().terms;
     String? currentTermId = StudentCourses().displayTermId;
 
     List<DropdownMenuItem<String>>? items;
     if (terms != null) {
       items = <DropdownMenuItem<String>>[];
-      for (CourseTerm term in terms) {
+      for (StudentCourseTerm term in terms) {
         items.add(DropdownMenuItem<String>(
           value: term.id,
           child: Text(term.name ?? '', style: getTermDropDownItemStyle(selected: term.id == currentTermId),)
@@ -198,8 +198,8 @@ class StudentCourseCard extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    String courseSchedule = _courseSchedule;
-    String courseLocation = _courseLocation;
+    String courseSchedule = course.section?.displaySchedule ?? '';
+    String courseLocation = course.section?.displayLocation ?? '';
     
 
     return ClipRRect(borderRadius: BorderRadius.all(Radius.circular(4)), child:
@@ -220,7 +220,7 @@ class StudentCourseCard extends StatelessWidget {
                     
                     Padding(padding: EdgeInsets.only(top: 6), child:
                       Row(children: [Expanded(child:
-                        Text('${course.shortName} (${course.number}) ${course.instructionMethod}', style: TextStyle(color: Styles().colors?.textBackground, fontFamily: Styles().fontFamilies?.medium, fontSize: 16),),
+                        Text(course.displayInfo, style: TextStyle(color: Styles().colors?.textBackground, fontFamily: Styles().fontFamilies?.medium, fontSize: 16),),
                       )]),
                     ),
                     
@@ -252,7 +252,7 @@ class StudentCourseCard extends StatelessWidget {
                               Image.asset('images/icon-location.png'),
                             ),
                             Expanded(child:
-                              Text(_courseLocation, style:
+                              Text(courseLocation, style:
                                 TextStyle(color: Styles().colors?.textBackground, fontFamily: Styles().fontFamilies?.medium, fontSize: 16,
                                   decoration: TextDecoration.underline, decorationColor: Styles().colors?.fillColorSecondary, decorationStyle: TextDecorationStyle.solid, decorationThickness: 1
                                 ),
@@ -273,27 +273,9 @@ class StudentCourseCard extends StatelessWidget {
 
   }
 
-  String get _courseSchedule {
-
-    String displayDays = course.section?.displayDays ?? '';
-    String displayTime = course.section?.displayTime ?? '';
-    if (displayDays.isNotEmpty) {
-      return displayTime.isNotEmpty ? "$displayDays $displayTime" : displayDays;
-    }
-    else {
-      return displayTime;
-    }
-  }
-
-  String get _courseLocation {
-    String buildingName = course.section?.buildingName ?? '';
-    String room = course.section?.room ?? '';
-    return (buildingName.isNotEmpty && room.isNotEmpty) ? "$buildingName $room" : buildingName;
-  }
-
   void _onLocaltion() {
     Analytics().logSelect(target: "Location Detail");
-    NativeCommunicator().launchMapDirections(jsonData: course.toMapsJson());
+    NativeCommunicator().launchMapDirections(jsonData: course.toJson());
   }
 
 }