+++ lib/ui/explore/ExploreDetailPanel.dart
@@ -17,6 +17,7 @@
 import 'package:flutter/material.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:illinois/model/Laundry.dart';
+import 'package:illinois/ui/athletics/AthleticsGameDetailPanel.dart';
 import 'package:illinois/ui/events/CompositeEventsDetailPanel.dart';
 import 'package:illinois/ui/laundry/LaundryRoomDetailPanel.dart';
 import 'package:rokwire_plugin/model/event.dart';
@@ -36,27 +37,36 @@ class ExploreDetailPanel extends StatelessWidget implements AnalyticsPageAttribu
 
   ExploreDetailPanel({this.explore, this.initialLocationData, this.browseGroupId});
 
+  static Widget? contentPanel({Explore? explore, Position? initialLocationData, String? browseGroupId}) {
+    if (explore is Dining) {
+      return ExploreDiningDetailPanel(dining: explore, initialLocationData: initialLocationData);
+    }
+    else if (explore is Event) {
+      if (explore.isGameEvent) {
+        return AthleticsGameDetailPanel(gameId: explore.speaker, sportName: explore.registrationLabel,);
+      }
+      else if (explore.isComposite) {
+        return CompositeEventsDetailPanel(parentEvent: explore);
+      }
+      else {
+        return ExploreEventDetailPanel(event: explore, initialLocationData: initialLocationData, browseGroupId: browseGroupId, );
+      }
+    }
+    else if (explore is LaundryRoom) {
+      return LaundryRoomDetailPanel(room: explore);
+    }
+    else { // Default for unexpected type
+      return null;
+    }
+  }
+
   @override
   Map<String, dynamic>? get analyticsPageAttributes => explore?.analyticsAttributes;
 
   @override
   Widget build(BuildContext context) {
-    if (explore is Dining) {
-      return ExploreDiningDetailPanel(dining: explore as Dining, initialLocationData: initialLocationData);
-    }
-    else if (explore is Event) {
-      Event event = explore as Event;
-      return event.isComposite ?
-        CompositeEventsDetailPanel(parentEvent: event) :
-        ExploreEventDetailPanel(event: event, initialLocationData: initialLocationData, browseGroupId: browseGroupId, );
-    }
-    else if (explore is LaundryRoom) {
-      return LaundryRoomDetailPanel(room: explore as LaundryRoom);
-    }
-    else { // Default for unexpected type
-      return Scaffold(
-        appBar: HeaderBar(),
-      );
-    }
+    return contentPanel(explore: explore, initialLocationData: initialLocationData, browseGroupId: browseGroupId) ?? Scaffold(
+      appBar: HeaderBar(),
+    );
   }
 }